Change to 4 spaces, fix small-para
small-para was not styling the second paragraph correctly.
There was a redundant font-weight style that was deleted.
The font-size style was being overwritten by the declaration in .para.
This was fixed by placing the small-para declaration after the para
declaration.

The proper fix would be to remove the "para" class from the second
paragraph in the html but we're not changing the html file in this
exercise.