Fix styling on confirm button
The confirm button had both "confirm" and "button" styles. This was a
problem since the "confirm" style was declared first and thus
overwritten. This was fixed by changing it to the "confirm-button"
id which was already on the button element.

There is a just a mess of classes everywhere in this html file since
this is a exercise I guess.