Finish flex exercise 5
I think I did this one into an acceptable fashion.

I think that the way that they did it in the solution was pretty good.
They made it that the header itself was a flexbox that contained the
text and the close button as the two flex items. They were then
separated using justify-contents: space-between;.

I should get to using those other values for justify-contents instead of
using margin: auto; for everything.